chore: converts to ts
