--- src/index.js
@@ -1,5 +1,0 @@
-import { createClient } from 'redis';
-
-createClient({
-  url: 'redis/saqib',
-});
