+++ src/public/index.html
@@ -19,5 +19,5 @@
       </div>
     </div>
   </body>
-  <script src="scripts.js"></script>
+  <script src="../dist/index.js"></script>
 </html>
